Remove :protocol option.
